+++ lib/ig_markets/format.rb
@@ -19,7 +19,7 @@ module IGMarkets
     # Returns a formatted string for the specified currency amount and currency. Two decimal places are used for all
     # currencies except the Japanese Yen.
     #
-    # @param [Float, Fixnum] amount The currency amount to format.
+    # @param [Float, Integer] amount The currency amount to format.
     # @param [String] currency The currency.
     #
     # @return [String] The formatted currency amount, e.g. `"USD -130.40"`, `"AUD 539.10"`, `"JPY 3560"`.
@@ -35,7 +35,7 @@ module IGMarkets
 
     # Returns a formatted string for the specified number of seconds in the format `[<hours>:]<minutes>:<seconds>`.
     #
-    # @param [Fixnum] value The number of seconds to format.
+    # @param [Integer] value The number of seconds to format.
     #
     # @return [String]
     def seconds(value)