+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.23 — Unreleased
 
-- 
+- Use `Integer` instead of `Fixnum` in preparation for Ruby 2.4
 
 ### 0.22 — October 3, 2016